Allow to select empty user or group
The rationale to introduce this change is as follow:
after 'both' a group and a user are selected, and an adequate
validation error is presented, it is not able to select "blank" in
either selection box - thus not able to fix the situation. Adding
`allow_blank` fixes this situation.

'prompt' was also removed since both `allow_blank` and `prompt` look
quite ugly, e.g:

```
Pick a user

user1
user2
```

By removing `prompt` the situation now looks as follow:

```

user1
user2
```