Set plgrid login in openid request for connected accounts
We want to avoid the situation when user connects his/hers account with
PLGrid using account `A`, start computation and after proxy expires
regenerate proxy certificate using PLGrid account `B`. In such a way we
will not be able to check computation status.

To solve this issue we are setting PLGrid user login in OpenId request
when vapor account is already connected with PLGrid.