Rename method name to better reflect its behavior
`User.with_active_computations` was renamed into
`User.with_submitted_computations`. The main purpose of this rename is
that active computation is also computation with `new` state (`new` means
that computation was created in our DB but not yet started on
infrastructure). On the other hand `with_submitted_computations` method
is used to trigger rimrock update task to update already submitted jobs.