Update automatic computation, run manual one
Now there is two modes of pipeline: automatic and manual. For automatic
all computations which can be started will be started in automatic way.
What is more once computation started it cannot be rerun or modified.
Manual pipeline allows to start computations many time and the order in
which computations start depends on the user action.