Fix problem with returning nils when data file type is not found
When synchronization is triggered and no data file type is found for the
path than data file is not created in DB. Unfortunately when data file
was not created in DB we were returning `nil`. This is a fix for this
problem, so instead of `[nil]` for unknown path `[]` is returned.